added GET stuff

+++ app/src/main/java/com/example/vehiclemaintenanceapplication/Backend/CarDataGet.kt
@@ -1,16 +1,16 @@
 package com.example.vehiclemaintenanceapplication.Backend
 
 import com.google.gson.annotations.SerializedName
-
+//car obd data get
 data class CarDataGet (
     @SerializedName("datetime") val datetime : String?,
     @SerializedName("speed") val speed : Double?,
     @SerializedName("engine_rpm") val rpm : Double?,
     @SerializedName("engine_runtime") val engineRuntime : Int?,
     @SerializedName("engine_coolant_temperature") val coolant : Double?,
-    @SerializedName("short_term_fuel_trim_1") val shortfueltrim1 : Int?,
+    @SerializedName("short_term_fuel_trim_1") val shortfueltrim1 : Double?,
     @SerializedName("short_term_fuel_trim_2") val shortfueltrim2 : Int?,
-    @SerializedName("long_term_fuel_trim_1") val longfueltrim1 : Int?,
+    @SerializedName("long_term_fuel_trim_1") val longfueltrim1 : Double?,
     @SerializedName("long_term_fuel_trim_2") val longfueltrim2 : Int?,
     @SerializedName("fuel_pressure") val fuelPressure : Int?,
     @SerializedName("engine_oil_temperature") val oilTemp : Int?,
